:recycle: refactor : STR-14 서브쿼리를 통한 comment size 구하기 수정
