+++ src/main/java/com/ssg/intern/dev/domain/bookmark/dao/BookmarkRepositoryCustomImpl.java
@@ -2,13 +2,16 @@ package com.ssg.intern.dev.domain.bookmark.dao;
 
 import com.querydsl.core.types.Order;
 import com.querydsl.core.types.OrderSpecifier;
+import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
+import com.ssg.intern.dev.domain.bookmark.entity.Bookmark;
 import com.ssg.intern.dev.domain.mypage.presentation.model.BookmarkProfileResponse;
 import com.ssg.intern.dev.domain.mypage.presentation.model.QBookmarkProfileResponse_Thumbnail;
 import com.ssg.intern.dev.global.SortingCondition;
 import lombok.RequiredArgsConstructor;
 
 import java.util.List;
+import java.util.Optional;
 
 import static com.ssg.intern.dev.domain.bookmark.entity.QBookmark.bookmark;
 import static com.ssg.intern.dev.domain.feed.entity.QFeed.feed;
@@ -31,6 +34,23 @@ public class BookmarkRepositoryCustomImpl implements BookmarkRepositoryCustom {
                 .where(bookmark.accountId.eq(accountId))
                 .orderBy(sort(sortingCondition))
                 .fetch();
+    }
+
+    @Override
+    public Optional<Bookmark> findBookmarkByFeedAndAccount(final long feedId, final long accountId) {
+        return Optional.ofNullable(queryFactory.selectFrom(bookmark)
+                .innerJoin(feed).fetchJoin()
+                .on(bookmark.feed.id.eq(feed.id))
+                .where(eqAccountId(accountId).and(eqFeedId(feedId)))
+                .fetchFirst());
+    }
+
+    private BooleanExpression eqFeedId(final long feedId) {
+        return bookmark.feed.id.eq(feedId);
+    }
+
+    private BooleanExpression eqAccountId(final long accountId) {
+        return bookmark.accountId.eq(accountId);
     }
 
     private OrderSpecifier<?> sort(final SortingCondition sortingCondition) {
