+++ src/main/java/com/ssg/intern/dev/domain/feed/dao/FeedRepositoryCustom.java
@@ -1,6 +1,7 @@
 package com.ssg.intern.dev.domain.feed.dao;
 
 import com.ssg.intern.dev.domain.feed.entity.Feed;
+import com.ssg.intern.dev.global.SortingCondition;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
