+++ src/main/java/com/ssg/intern/dev/domain/mypage/presentation/MypageApi.java
@@ -1,5 +1,6 @@
 package com.ssg.intern.dev.domain.mypage.presentation;
 
+import com.ssg.intern.dev.domain.feed.presentation.model.MyReviewProfileResponse;
 import com.ssg.intern.dev.domain.mypage.presentation.model.BookmarkProfileResponse;
 import com.ssg.intern.dev.domain.mypage.service.MypageQueryService;
 import com.ssg.intern.dev.domain.mypage.service.MypageCommandService;
@@ -17,7 +18,6 @@ public class MypageApi {
 
     private final MypageQueryService mypageQueryService;
     private final MypageCommandService mypageCommandService;
-    private final MypageQueryService mypageQueryService;
 
     @GetMapping("/me/thumbnails")
     public BookmarkProfileResponse getThumbnails(@RequestParam String sortingCondition,
